backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
Babel==2.14.0
bcrypt==4.2.1
beanie==1.22.4
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
boto3==1.34.38
botocore==1.34.162
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.9.0
flake8==7.0.0
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.4.0
googlemaps==4.10.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
idna==3.10
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy-model==0.2.0
loguru==0.7.2
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
motor==3.5.3
mpmath==1.3.0
mypy==1.7.1
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0
notebook==7.2.1
notebook_shim==0.2.3
numpy==1.26.3
openai==1.2.3
opencv-python==4.10.0.84
orjson==3.10.15
overrides==7.6.0
packaging==24.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
pipdeptree==2.25.0
platformdirs==4.3.6
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.10.2
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.10.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.4
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.35.1
sympy==1.12
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.27.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.11
win32_setctime==1.2.0